+++ server/schedule/priorityscheduling/priority_schedule.go
@@ -19,44 +19,3 @@ type HighPrioritySchedule interface {
 	GetHighPriorityScheduleRegions() []uint64
 	RemoveHighPriorityScheduleRegion(id uint64)
 }
-
-// MockHighPrioritySchedule provide a mock HighPrioritySchedule, only used for unit test
-type MockHighPrioritySchedule struct {
-	ids map[uint64]struct{}
-}
-
-// NewMockHighPrioritySchedule return a mock HighPrioritySchedule
-func NewMockHighPrioritySchedule() *MockHighPrioritySchedule {
-	return &MockHighPrioritySchedule{
-		ids: map[uint64]struct{}{},
-	}
-}
-
-// AddHighPriorityScheduleRegions implements the HighPrioritySchedule.AddHighPriorityScheduleRegions
-func (mh *MockHighPrioritySchedule) AddHighPriorityScheduleRegions(regionsIDs ...uint64) {
-	for _, id := range regionsIDs {
-		mh.ids[id] = struct{}{}
-	}
-}
-
-// GetHighPriorityScheduleRegions implements the HighPrioritySchedule.GetHighPriorityScheduleRegions
-func (mh *MockHighPrioritySchedule) GetHighPriorityScheduleRegions() []uint64 {
-	if len(mh.ids) < 1 {
-		return nil
-	}
-	returned := make([]uint64, 0, len(mh.ids))
-	for id := range mh.ids {
-		returned = append(returned, id)
-	}
-	return returned
-}
-
-// RemoveHighPriorityScheduleRegion implements the HighPrioritySchedule.RemoveHighPriorityScheduleRegion
-func (mh *MockHighPrioritySchedule) RemoveHighPriorityScheduleRegion(id uint64) {
-	delete(mh.ids, id)
-}
-
-// Reset is only used for unit test
-func (mh *MockHighPrioritySchedule) Reset() {
-	mh.ids = map[uint64]struct{}{}
-}
